NimBLEL2CAPChannel: Preliminary handling out of-out-of-memory situations
When `ble_l2cap_send` returns -ENOMEM, then the internal buffers have not
yet been transmitted to the l2cap peer. Since we have no way to find out
when this happened (unfortunately they are not using the same mechanism
as for the 'transmit stalled'), we discard our buffers, wait and retry.

NOTE: This needs a check whether the connection is still alive.